requirements.txt
altgraph==0.17
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
distlib==0.3.2
Django==3.2.5
django-cors-headers==3.7.0
django-simple-chatbot==0.0.9
djangorestframework==3.12.4
fastapi==0.68.1
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
h11==0.12.0
idna==2.10
image==1.5.33
jieba3k==0.35.1
joblib==1.0.1
lxml==4.6.3
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.19.1
oauthlib==3.1.0
pandas==1.3.2
pefile==2021.5.24
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.2.0
pydantic==1.8.2
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
regex==2021.7.6
requests==2.25.0
requests-file==1.5.1
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
starlette==0.14.2
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.0
tweepy==3.9.0
typing-extensions==3.10.0.2
unicorn==1.0.3
urllib3==1.26.2
uvicorn==0.15.0
virtualenv==20.7.0
virtualenv-clone==0.5.6
XlsxWriter==3.0.1